Added Lines of code plt.savefig ( ) and plt.grid ( )
We added lines of code to save the T-x-y plot using plt.savefig( ), and plt.grid( ) to add grid lines.
We tried to calculate the saturated vapor pressure of pentane at 300K by obtaining the Antoine's coefficient at the high and  low temperature range.
We added a descriptive comment on the Readme.md file.
We added a line of comment to define Ps in the raoult_law_kvalue.py.